requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
langchain-ollama
chromadb
rank_bm25
duckduckgo-search
pyyaml
python-dotenv
rich
tqdm
click
ddgs